+++ ChangeLog.md
@@ -8,6 +8,7 @@
 
     - Add 'stdinFromNull' option to TTP::commandExec(), defaulting to true
     - TTP::commandByOs() is renamed to commandByOS() for consistency reasons (todo #96)
+    - TTP::RunnerCommand module is renamed RunnerVerb (todo #93)
 
 ### 4.5.1